requirements.txt
flask>=2.3
pyserial>=3.5
